extract contract to its own class
